requirements.txt
Flask>=2.0.1
Flask-SocketIO>=5.3.0
opencv-python>=4.8.1.78
PyAudio>=0.2.13
numpy>=1.26.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
PyQt5>=5.15.0
python-socketio[client]>=5.3.0
requests>=2.28.0
face_recognition>=1.3.0
